refactor: Title to shared
