+++ src/features/about-video/about-video.tsx
@@ -1,4 +1,4 @@
-import { Title } from '@/app/components';
+import { Title } from '@/shared/ui/title';
 
 import './about-video.scss';
