+++ src/features/course-main/course-main.tsx
@@ -1,6 +1,6 @@
 import { useLoaderData } from 'react-router-dom';
 import { getCourseStatus } from './utils';
-import { DateLang, SectionLabel, Title } from '@/app/components';
+import { DateLang, SectionLabel } from '@/app/components';
 import { useTitle } from '@/app/hooks';
 import { selectCourse } from '@/app/hooks/use-course-by-title/utils/select-course.ts';
 import { Course } from '@/app/types';
@@ -8,6 +8,7 @@ import Image from '@/features/image';
 import { ArrowIcon } from '@/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { Subtitle } from '@/shared/ui/subtitle';
+import { Title } from '@/shared/ui/title';
 
 import styles from './course-main.module.scss';
