+++ src/features/home/about/about.tsx
@@ -1,9 +1,8 @@
-import { Title } from '@/app/components';
-
 import image from '@/assets/about.webp';
 import Image from '@/features/image';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { Subtitle } from '@/shared/ui/subtitle';
+import { Title } from '@/shared/ui/title';
 
 import './about.scss';
