+++ src/features/home/numbers/numbers.tsx
@@ -1,8 +1,8 @@
-import { Title } from '@/app/components';
 import image from '@/assets/map.webp';
 import Image from '@/features/image';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { Subtitle } from '@/shared/ui/subtitle';
+import { Title } from '@/shared/ui/title';
 
 import './numbers.scss';
