+++ src/features/home/speakers/speakers.tsx
@@ -1,9 +1,8 @@
-import { Title, TitleType } from '@/app/components';
-
 import image from '@/assets/speakers-wanted.webp';
 import Image from '@/features/image';
 import { EmailIcon } from '@/icons';
 import { Subtitle } from '@/shared/ui/subtitle';
+import { Title, TitleType } from '@/shared/ui/title';
 
 import './speakers.scss';
