+++ src/features/home/support/support.tsx
@@ -1,11 +1,10 @@
-import { Title, TitleType } from '@/app/components';
-
 import { LINKS } from '@/app/const';
 import image from '@/assets/support.webp';
 import Image from '@/features/image';
 import { ArrowIcon } from '@/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { Subtitle } from '@/shared/ui/subtitle';
+import { Title, TitleType } from '@/shared/ui/title';
 
 import './support.scss';
