+++ src/features/mentors-wanted/mentors-wanted.tsx
@@ -1,10 +1,10 @@
 import classNames from 'classnames';
 import classNamesBind from 'classnames/bind';
 import { Link } from 'react-router-dom';
-import { Title } from '@/app/components';
 import mentorImg from '@/assets/mentors-wanted-poster.webp';
 import Image from '@/features/image';
 import { Paragraph } from '@/shared/ui/paragraph';
+import { Title } from '@/shared/ui/title';
 
 import styles from './mentors-wanted.module.scss';
