+++ src/features/school/study-with-us/study-with-us.tsx
@@ -1,10 +1,11 @@
 import { studyOptions } from './constants';
-import { OptionItem, SectionLabel, Title } from '@/app/components';
+import { OptionItem, SectionLabel } from '@/app/components';
 
 import image from '@/assets/rs-school.webp';
 import Image from '@/features/image';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { Subtitle } from '@/shared/ui/subtitle';
+import { Title } from '@/shared/ui/title';
 
 import './study-with-us.scss';
