+++ src/features/trainers/trainers.tsx
@@ -1,6 +1,6 @@
 import { TrainerCard } from './trainer-card';
-import { Title } from '@/app/components';
 import { Trainer } from '@/features/trainers/trainers.types.ts';
+import { Title } from '@/shared/ui/title';
 
 import './trainers.scss';
