+++ src/shared/ui/title/title.scss
